Introducing JoseConfiguration and passing it to factories instead of JoseFraction
